Add printing of all values in loaded CSV file
In parseCSV.py the CSV file "All Data.csv" was loaded using the python
csv library and the values for every row in the csv were printed to test
that all of the values were read properly.

The ability to load the csv file into python is necessary to manipulate
the csv into a format that can be easily imported into a mysql database.