Refactor ERD to reduce data duplication
The entity relationship diagram for the database was updated to include
foreign keys in the accident table for the driver, condition and
location tables. The foreign key reference to the accident table was
removed in the driver, condition and location tables.

It was necessary to move the foreign keys into the accident table as
the set of locations, conditions, and drivers is less than the amount
of accidents that occur. In the condition table the dataset only
includes 25 rows and if accidentID remained a foreign key the condID
would have to repeat to store a condition for every accident.