Add reset index to parseCSV to correct ID values
The reset index function was used after every drop duplicates call
to reset the index to be used for table IDs so that the IDs are
incremental integer values.

Reset index was necessary to output the correct ID for each table
as before after the duplicate rows are droppped the index values stay
the same as what was in the original table and are not incremental.