Implement separation of date and time to accident
The value of date and time was separated into separate columns and the
redundant date and time column was removed. The columns were renamed to
the same names as the database and exported to accident.csv.

Date and time was separated into 2 separate columns to align with the
database design and the columns were renamed so that the csv can be
easily imported into the database.