Add foreign keys to the accident CSV
The merge function was used to merge location, weather, and driver with
the accident table so that foreign key values could be set in the CSV.
Three new columns were defined driverID, condID and locID that stores
the values that represent the foreign keys.

Adding the foreign keys that come from the driver, weather and location
tables is necessary for proper references between tables.