Fix error vehicle not stored after dropping nulls
The df_export_vehicle dataframe was not set equal to the value of
df_export_vehicle.dropna and as a result the csv that was generated
was not updated properly to reflect the intended behavior.

The df_export_vehicle has to be updated properly for when the dataframe
is exported to a csv.