Fix accDate type to date and excel error
Previously the date would not properly import in the database with the
type of accDate as date. This error occurred because opening the file
in excel and saving it changed the date format to the setting in excel
changing the format from YYYY-MM-DD to DD-MM-YYYY which was not accepted
by the database. The type of accDate was previously changed to varChar
and this change was reverted and the csv with the proper date formatting
was imported.

The accDate field holds the date and it is better that it is stored
as a date data type.