Fix error where format of time field was wrong
The beginning of the time field was being removed when attempting to
remove the unnecessary characters at the end of the field. Before the
string was split by the 'Z' at the end as a delimiter and this resulted
in the beginning of the time being removed. This was fixed by splitting
the field by '.' as a delimiter to remove the useless end of the string.

The time field was formatted properly so that the data in the database
is complete.