Add new city dataframe to generate city CSV
A new dataframe city is used to store information that will be in the
city table. This dataframe is set equal to the city data field in
the location dataframe and the locID of each city. The accID field
was also added to the script so that the accID is equal to the index
export_accident so no additional manipulation is needed before importing
the csv files.

It is necessary to add a city table to follow the change of design so
that the database is in 3NF. The city table removes a transitive
dependency between city and streetAddress. It is also useful to refactor
certain parts of the script so that the output CSV is correct and can
be immediately imported into the database.